+++ packages/common/src/cis2/index.ts
@@ -1,2 +1,3 @@
 export type { CIS2 } from './util';
+export { tokenAddressFromBase58, tokenAddressToBase58 } from './util';
 export * from './CIS2Contract';